@@ -5,12 +5,13 @@
   "private": true,
   "license": "UNLICENSED",
   "author": "Liquid Labs, LLC",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+ssh://git@github.com/Liquid-Labs/intraweb.git"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=16.0.0"
   },
   "scripts": {
     "start": "node app.js"
@@ -19,6 +20,16 @@
     "@google-cloud/storage": "^5.8.5",
     "express": "^4.17.1",
     "express-async-handler": "^1.1.4",
-    "marked": "^2.1.3"
+    "gcp-metadata": "^4.3.0",
+    "google-auth-library": "^7.7.0"
+  },
+  "devDependencies": {
+    "@liquid-labs/catalyst-scripts": "^1.0.0-alpha.44",
+    "marked": "^3.0.2"
+  },
+  "liq": {
+    "eslint": {
+      "target": "app.js lib/**/*.js"
+    }
   }
 }
